refactor(web): share day/week interaction adapter helpers
Lift identical smart-scroll, ownership, cursor, and moved-predicate
helpers into @web/grid/interaction, and extract Day commit/geometry
modules to match Week's adapter layout without merging the adapters.
