+++ packages/web/src/grid/interaction/commit/timed-moved.ts
@@ -1,0 +1,11 @@
+import { type TimedDragVisual } from "@web/grid/interaction/types/timed-drag.types";
+import { type TimedResizeVisual } from "@web/grid/interaction/types/timed-resize.types";
+
+export const hasTimedDragVisualMoved = (visual: TimedDragVisual) =>
+  visual.dayDate !== visual.initialDayDate ||
+  visual.startMinutes !== visual.initialStartMinutes ||
+  visual.endMinutes !== visual.initialEndMinutes;
+
+export const hasTimedResizeVisualMoved = (visual: TimedResizeVisual) =>
+  visual.startMinutes !== visual.initialStartMinutes ||
+  visual.endMinutes !== visual.initialEndMinutes;
